@@ -3,8 +3,9 @@ package com.weidian.lib.hera.utils;
 import android.app.Activity;
 import android.content.Context;
 import android.os.Build;
-import android.support.annotation.ColorInt;
-import android.support.annotation.IntRange;
+import androidx.annotation.ColorInt;
+
+import androidx.annotation.IntRange;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.WindowManager;
