@@ -3,8 +3,10 @@ package com.weidian.lib.hera.sample;
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+
 import android.view.View;
+
+import androidx.annotation.Nullable;
 
 public class ForResultActivity extends Activity {
 
